Start adding sampling op
Added an op to sample from a normal distribution with any shaped mean
and stddev input. The batch dimension is always dimension 0.

Started adding the Sampel() and Rsample() methods to Normal, then
realized that to be consistent witht the aforementioned sampling
procedure, the Normal should take any input and use the batch dimension
as dim 0. First, change the normal to use batch dim as dim 0, then
implement the Rsample() method.