Input any size mean/stddev to normal
Before the Normal only worked with scalars or vectors. Now, it works
with any sized input tensor for mean and stddev and just makes a bunch
of element-wise normal distributions. For any input, dim = 0 is
considered the batch dimension, except if the input has the same shape
as the normal's shape, in which case we consider a single sample in the
batch.