Create Quantiler interface for Cdfinv method
Created a new interface `Quantiler` which calculates the Quantile
function of a node. The `Cdfinv` method was renamed to `Quantile`. This
was required for `*IID` to implement `Distribution`, since the qantile
function of a distribution of i.i.d. random variables does not exists.
For any x, the quantile function of x is non-unique for a distribution
of i.i.d. random variables.